Updated documentation for StanwoodAnalytics class.
