+++ .jazzy.yaml
@@ -2,15 +2,19 @@ output: docs
 clean: true
 skip_undocumented: true
 sdk: iphone
-swift_version: 4.1.0
-module_version: 0.0.53
-xcodebuild_arguments: [clean,build,-project,_Pods.xcodeproj,-scheme,StanwoodAnalytics]
-source_directory: StanwoodAnalytics/Classes
 author: Ronan O Ciosoig
-author_url: ronan.o.ciosoig@stanwood.io
-copyright: 2018 Stanwood GmbH
+author_url: https://www.stanwood.io 
+copyright: '© 2018 [Stanwood GmbH] (https://www.stanwood.io) under MIT license.'
 readme: README.md
 podspec: StanwoodAnalytics.podspec
 github_url: https://github.com/stanwood/Stanwood_Analytics_iOS
 min_acl: public
 theme: fullwidth
+custom_categories: 
+  - name: Core
+    children:
+    - AlertFactory
+    - DataStore
+    - StanwoodAnalytics
+    - Tracker
+    - TrackerKeys
