+++ docs/search.json
@@ -1,1 +1,1 @@
-{}
+{"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV9eventNameSSvp":{"name":"eventName","abstract":"<p>The event name</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV6itemIdSSSgvp":{"name":"itemId","abstract":"<p>Item Id</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV4nameSSSgvp":{"name":"name","abstract":"<p>Name</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV11descriptionSSSgvp":{"name":"description","abstract":"<p>Description</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV8categorySSSgvp":{"name":"category","abstract":"<p>Category</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV11contentTypeSSSgvp":{"name":"contentType","abstract":"<p>Content Type</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersV06customD0s10DictionaryVySSypGvp":{"name":"customParameters","abstract":"<p>Define custom parameters here</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersVACSS9eventName_tcfc":{"name":"init(eventName:)","abstract":"<p>Init with event name only. The remaining parameters are all set to nil.</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersVACSS9eventName_SSSg6itemIdAE4nameAE11descriptionAE8categoryAE11contentTypetcfc":{"name":"init(eventName:itemId:name:description:category:contentType:)","abstract":"<p>Init with all the parameters. Event name is required, the remaining parameters are optionals.</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersVACSS9eventName_SSSg11contentTypetcfc":{"name":"init(eventName:contentType:)","abstract":"<p>Init with event name and content type only. The remaining parameters are all set to nil.</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html#/s:17StanwoodAnalytics18TrackingParametersVACSS9eventName_SSSg4nametcfc":{"name":"init(eventName:name:)","abstract":"<p>Init with event name and name only. The remaining parameters are all set to nil.</p>","parent_name":"TrackingParameters"},"Structs/TrackingParameters.html":{"name":"TrackingParameters","abstract":"<p>A struct to hold all the values to be sent to the trackers.</p>"},"Other Protocols.html#/s:17StanwoodAnalytics15ParameterMapperP":{"name":"ParameterMapper","abstract":"<p>Protocol to map tracking parameters to a dictionary of [String:NSString] and removes all optionals.</p>"},"Other Enums.html#/s:17StanwoodAnalytics13TrackingEventO":{"name":"TrackingEvent","abstract":"<p>An enumeration to define the event type.</p>"},"Classes/StanwoodAnalytics/Builder.html#/s:17StanwoodAnalyticsAAC7BuilderC3addAdA7TrackerC7tracker_tF":{"name":"add(tracker:)","abstract":"<p>Add a tracker to the builder.</p>","parent_name":"Builder"},"Classes/StanwoodAnalytics/Builder.html#/s:17StanwoodAnalyticsAAC7BuilderC23setNotificationDelegateADSo16UIViewControllerC8delegate_tF":{"name":"setNotificationDelegate(delegate:)","abstract":"<p>Set a delegate to display local notifications. This is used for debugging the tracking. It will display a local notification for each time track is called.</p>","parent_name":"Builder"},"Classes/StanwoodAnalytics/Builder.html#/s:17StanwoodAnalyticsAAC7BuilderC5buildAByF":{"name":"build()","abstract":"<p>Build the analytics object.</p>","parent_name":"Builder"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC4KeysV":{"name":"Keys","abstract":"<p>Keys for user parameters.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC13TrackingEventO":{"name":"TrackingEvent","abstract":"<p>The event types used for tracking.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAACA2B7BuilderC7builder_tcfc":{"name":"init(builder:)","abstract":"<p>Init method using the Builder pattern.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC7builderAB7BuilderCyFZ":{"name":"builder()","abstract":"<p>The Builder for this class.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC5trackyAA18TrackingParametersV08trackingE0_tF":{"name":"track(trackingParameters:)","abstract":"<p>Track the parameters. Which parameters that are actually tracked depends on","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC15trackingEnabledSbyFZ":{"name":"trackingEnabled()","abstract":"<p>Returns the stored value for the tracking setting.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC11setTrackingySb6enable_So16UIViewControllerCSg2ontF":{"name":"setTracking(enable:on:)","abstract":"<p>Set this value to false to stop the tracking, and on the next start it will be fully disabled.","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC5trackyAA11TrackerKeysV07trackerE0_tF":{"name":"track(trackerKeys:)","abstract":"<p>Track specific keys.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC5trackySo7NSErrorC5error_tF":{"name":"track(error:)","abstract":"<p>Track NSError objects.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html#/s:17StanwoodAnalyticsAAC11trackScreenySS4name_SSSg9classNametF":{"name":"trackScreen(name:className:)","abstract":"<p>Helper function to track the screen name along with the class name.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics/Builder.html":{"name":"Builder","abstract":"<p>The builder for this class.</p>","parent_name":"StanwoodAnalytics"},"Classes/StanwoodAnalytics.html":{"name":"StanwoodAnalytics","abstract":"<p>The base class for Analytics</p>"},"Core.html":{"name":"Core"},"Other Enums.html":{"name":"Other Enumerations","abstract":"<p>The following enumerations are available globally.</p>"},"Other Protocols.html":{"name":"Other Protocols","abstract":"<p>The following protocols are available globally.</p>"},"Other Structs.html":{"name":"Other Structures","abstract":"<p>The following structures are available globally.</p>"}}
